Rename legacy exposition to twisted exposition to better reflect usage of remaining code
